Removal of constraint inputs to poolPlan call, improving compile time
Summary:
This affects the order in which variables are evaluated in the pool plan, and was observed to decrease the number of
calls to the cycle estimator.  Although those calls are cached, this makes a large difference in compile time.

Also, minor tidy to cycle estimate function

Fixes T29219

Test Plan: CI

Reviewers: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T29219

Differential Revision: https://phabricator.sourcevertex.net/D35751