Correctly avoid the slow path when selecting vectorInner elementwise ops
Summary:
Correction - if vectorInner operates by identifying a `data` vector and adds/multiplies by a vector `B` that is repeated N = dataLen/BLen times then:
There is a slow path where `B` is not a multiple of the vector width of HALF for the multiply, add (and subtract==add) vertices.
The code was picking a`data` being a multiple of the vector width so for example with BLen=2 the slow path would be picked with even `N`.
Also, the newly added DIVIDE wasn't taken into account.  There is no similar super slow path to be obviously avoided

Discovered debugging task Ref T22299

Reviewers: tomm, claudiop, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: claudiop, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T22299

Differential Revision: https://phabricator.sourcevertex.net/D36324