Corrections to the exchange data size for pooling plan - including large difference for maxPool Bwd
Summary:
The equation for bytes exchanged was wrong before, and the calculation ignored the fact that some vertices have more inputs than others.

Reviewed and checked the bytes needed to make a given partition of the output against what is actually attached to the codelets and the amount of data exchanged. This was all consistent - seems logical too.
In theory some types of vertices are connected to >2x the data - in particular max pool grad.  Evidence showed this to be more like 2x the data being exchanged, which maybe just the mapping from the test.  Written assuming all data required as an input to all vertices will be exchanged.

Ref T29219

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T29219

Differential Revision: https://phabricator.sourcevertex.net/D35989