Add test to verify that the compiler generates RPT loops.
Summary:
Popops codelets tests are added to verify that certain vertices take the
expected number of cycles (within a tolerance) when executing on Sim1.
These vertices/operation/data-type combinations are written in C++ and we know
that POPC generates RPT inner loops for them (instead of BRNZDEC loops) for
Sim1.

The test is done by using the UnaryCodeletsTest/BinaryCodeletsTest executables
that have an option to compare execution cycles when a vertex is run on the
simulator with those reported when run on IpuModel (i.e. the cycle estimators).

Cycle estimators for these vertices are modified to a more precise value.

The UnaryCodeletsTest/BinaryCodeletsTest executables are modified to support a
configurable threshold for the difference in cycles to be considered as an
error.

The tests are not run for Sim2 because currently POPC does not generate RPT
loop for it.

Test Plan: Run updated poplibs regression

Reviewers: timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: davidb, janekvo

Maniphest Tasks: T15543

Differential Revision: https://phabricator.sourcevertex.net/D36420