Allow fwd pass to be done by dense operation
Summary:
Allow the fwd pass to be done as a dense operation. This can only be hit if
the plan is constrained to use the dense version.

Implemented by having the SparseTensors nzValues being the whole dense tensor and the
meta info being invalid. The partioner will either initialise the buckets for sparse
or the dense tensor for dense version.

Test Plan: included

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Subscribers: godfrey.da.costa

Maniphest Tasks: T24264

Differential Revision: https://phabricator.sourcevertex.net/D36880