Use FP control to force saturation mode in block sparse Softmax
Summary:
There is a high probability of overflow when the maximum
computed in the stabilisation of each row is positive and
minimum values are inserted due to sparsity/masking.

Force saturation around the stabilisation of the softmax
so that NaNs on overflow don't occur even if such
mode is set.

Ref T31131

Test Plan: Added new tests and changed the way random numbers are generated

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Maniphest Tasks: T31131

Differential Revision: https://phabricator.sourcevertex.net/D37532