Test model for CTC Inference, and validation of CTC Loss and Inference
Summary:
Note - the model for CTC Loss was already landed, but is called by the tool in this change.

Create model for CTC  Inference which is easy to understand as a reference implementation.  A beam search and a brute force "all paths" option are provided, however "all paths" is only practical for very small inputs even on a server.
The tool provided can be used with tensorflow to validate CTC Inference.  That requires tensorflow to be installed.  Automated tests can be run of reasonable size.
The tool can be used to validate CTC Loss with pytorch.  This is more manual (could be improved) but as the algorithm is far simpler the testing this allows seems sufficient.

Ref T31895, T31896

Test Plan: Validate against pytorch/tensorflow implementations

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Subscribers: samk, godfrey.da.costa, saml

Maniphest Tasks: T31896, T31895

Differential Revision: https://phabricator.sourcevertex.net/D36817