Add Sim2 to Unary/BinaryOp SupervisorCycleTest
Summary:
The tests verifying that the compiler emits RPT loops where enabled only for
Sim1. Now popc generates RPT loops for Ipu2 as well, so we add Sim2 to the
tests in CMakeList.txt

Handling of command line parameters in the two C++ test executables is slighlty
modified so that the appropriate IpuModelX can be automatically matched to the
specified SimX device, for ease of use of the command line.

Also the INVERSE, RSQRT and SQRT vertices *for HALF data type* are removed from
the test (they were included by mistake in the first place, as RPT loops have
never been generated for these operators/data type).

Additionally, the test executables are run through a python script which, in
case of failure, prints an additional message prompting the tester to check the
values of the arcane cmake settings required to convince the compiler to emit
RPT loops.

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp, Eddy, mattl

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, mattl

Maniphest Tasks: T31752

Differential Revision: https://phabricator.sourcevertex.net/D38816