Adding the operation log add to the reduction library
Summary:
A new operation defined as result = a+std::log(1+std::exp(b-a)); (Where a > b - chosen while carrying out the operation to improve accuracy).

This is the required operation when the data to reduce is log probabilities.
The operation is needed in CTC loss implementation and may be needed in other instances.
Added with C++ codelets, but able to access all specialisations.
Defined update as implementing another "log add" when combing with the output.
Define applying scale as the operation scaledResult=result + scale;  Consistent with log scaled arithmetic.

The reduce_op test can be used to excersise this fully.
The ReductionTests test (With --test=Ops) can test this operation but doesn't support update or scale (for any operation)

Fixes T32996

Test Plan: Added several tests to excersise the operation with update,scale, and the codelet implementation in the 4 places it needed modifying.

Reviewers: saml, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T32996

Differential Revision: https://phabricator.sourcevertex.net/D39086