Simplification to CTC Loss by initialising temporary memory, and use of log::probabilityZero and log::probabilityOne
Summary:
A tidy up step to change from using `log::min`  and `0` to `log::probabilityZero` and `log:;probabilityOne`.

The main change is to initialise the temporary inputs to all the vertices to `log::probabilityZero` at the start of execution, instead of doing it as we go.  This is both the inputs around the edge of the partitioned [t,el] shaped working array and those that are exchanged between tiles (between partitions of the [t,el] working array).
The benefits are that we avoid adding extra small copy programs when constructing the program. Plus the internal spaces being initialised means that the vertex code doesn't need to copy initialiser to initialiser when it has nothing to do.

Additionally, by extending the size of the tempLabels data by 1 timestep (1 element for alpha, 2 for beta) the logic to copy but shuffle along by 1 timestep is simplified.  The same amount of data is exchanged/copied, just there si no need to insert and extra zeroed timestep of data

Ref T32938

Test Plan: CI

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T32938

Differential Revision: https://phabricator.sourcevertex.net/D39645