Ignore partial types that are smaller than the output type
Summary:
9235f68393 introduced an bug where for a joint plan no vertex candidates
were found because the partial type was smaller than the output type and
we hit a segfault. This happened to work before but only out of chance -
there were no tests to covering this case. Fix this bug by ignoring the
partial type if it is smaller than the output type.

If you have a mixture of operations with float and half outputs this lets
you globally set the partial type option and have it only apply to the
operations with half outputs.

Ref T33038

Reviewers: samk, godfrey.da.costa, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: samk, godfrey.da.costa, #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: danielj, dominicm, marieanne

Maniphest Tasks: T33038

Differential Revision: https://phabricator.sourcevertex.net/D39200