Add non-inplace variants for nonlinearities
Summary:
Third attempt at this:
The existing optimized assembly vertices for TANH, SIGMOID, RELU were handling
only the in-place variants.
This change adds entry points for the not-in-place variants, using the existing
assembly code for the inner loops.

The NonLinearity popnn code is modified so that for those operations we use the
new vertices, instead of doing a copy of the tensor and calling the in-place
version.

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm, timp

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Subscribers: timp

Maniphest Tasks: T29651

Differential Revision: https://phabricator.sourcevertex.net/D41131