Broadcast and map misc small tensors in CTC Loss to avoid repeated copy and broadcast
Summary:
The labelLength and dataLength single elements were getting exchanged/broadcast to the tiles where they are needed on every compute step.  Although only 100 cycles, when work gets highly fragmented by the planner this becomes significant.

Likewise the labels data itself.  To stop unnecessary exchange we need to map `previous, a,b,c,... next` where the vertex processes a result of -a,-b,-c but relies on the data and symbol for previous (when finding alpha) OR next (when finding beta).  So a kind of sliding window 2 greater than is needed is produced, capturing all that any vertex on this tile is needed.
Alpha connects to `previous,a,b,c...` and Beta to `a,b,c,...,next`

This made it simpler and more logical to attach the vertex to a single contiguous series of labels, instead of `previous` and `a,b,c,...` as 2 inputs.  It felt like the vertex code could be simplified as a result, which worked out for vertices calculating alpha, but not so obvious for beta.

Ref T32938

Test Plan:
CI.
Observed no longer any unnecessary exchange per compute step

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T32938

Differential Revision: https://phabricator.sourcevertex.net/D39989