Tidy/simplify vertical wave tensor implementation of CTC Loss
Summary:
Reduce the size of temporary label slice tensor to 1 element  per tile (2 for beta), as this is all that is needed.  Remove unnecessary copies of alpha, beta into the temporary input's timeslice.

All of this simplifies this greatly and the results are faster, but with unwanted copies when partitioning >1 timestep per tile.  This is because we explicitly ask for the copies though!  The overhead isn't massive and although those copies could be avoided, a loop implementation is necessary and will affect the way the copies are avoided, so they are left along for now.

Ref T32938

Test Plan: CI

Reviewers: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #poplar_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T32938

Differential Revision: https://phabricator.sourcevertex.net/D40258