Account for interleaved memory in Cast vertex cycle estimates
Summary:
I've changed the common castWorkerCycles function which covers Cast, CastSupervisor, and Cast2D vertices. The only change is to account for the number of read ports in the calculation for the load cycles.

The only case where this has any impact on the cycle estimate is in the Cast from float to half, as we can utilise 128 bit reads and 64 bit writes to reduce the inner loop cycles by 1.

Test Plan: CI

Reviewers: tomm, samk, claudiop, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T31467

Differential Revision: https://phabricator.sourcevertex.net/D40627