Map CTC Loss gradient and input tensors more carefully
Summary:
Mapping did split the innermost dimension which resulted in increased exchange memory.
The input mapping was the primary culprit for this, as I expect output reduction made sensible choices.
Taking control of the input mapping had a large effect on exchange code, that of the softmax and reduced gradient output less so,
but it seemed good to be consistent.

Test Plan:
CI, verified improvements manually

Ref T32938

Reviewers: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T32938

Differential Revision: https://phabricator.sourcevertex.net/D40747