Tidy CTC Loss codelets
Summary:
Try to simplify and make each codelet as similar as possible.
Lots of little changes in the codelet code but some slight changes to how it all works:

1. We can avoid a logAdd in the maths by conditionally using the sum from the blank in an adjacent symbol.  This is easy to change in the alpha case, but the compute ordering was wrong for beta. Calculating Beta in reverse order makes this possible
I'm not sure if in reality (with a highly split el, or 1 output per worker in a supervisor vertex) that this will translate to faster, but it's certainly more efficient.
2. The count down for vertices that calculate beta is now one less.  This makes dealing with the count more similar in all vertices
3. Initialising the 1st beta timeslice to zero avoids the need for the vertex to do the initialising.

Ref T32938

Test Plan: CI

Reviewers: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T32938

Differential Revision: https://phabricator.sourcevertex.net/D40723