Outline matmuls in triangular solver with properly layed out arguments
Summary:
We try to keep block elements together, but this layout is not good for block-combining matmuls.
This commit outlines and reuses grouped matmuls used by triangular solver.
Due to divide-and-conquer nature of the blocking algorithm, the effect is getting more noticeable each split.
As a result we have 30% memory reduction in Gaussian Process PoC.

Ref T34392.

Test Plan: CI

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, samk

Maniphest Tasks: T34392

Differential Revision: https://phabricator.sourcevertex.net/D40892