Account for interleaved memory in ReduceAdd vertex cycle estimates
Summary: Ref T31467. In the ReduceAdd vertex cycle estimates, there is a parameter that is provided to the cycle estimator function that tells us whether we have constrained partials. This is always provided as false currently. If we force the use of interleaved cycle estimates, all we need to do is set this bool to true.

Test Plan: CI

Reviewers: tomm, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T31467

Differential Revision: https://phabricator.sourcevertex.net/D40847