Optimise CTC Loss codelets to avoid unnecessary log-add to values that are -inf
Summary:
The first time a codelet writes to a gradient output the value will always be -inf + newValue.  There is no point in doing the addition as the result will be newValue.
The fact that the addition wasn't needed has always been the case, but when each tile had a bigger slice of the labels input the amount of wasted effort would have been less.

Ref T31901

Test Plan: CI

Reviewers: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T31901

Differential Revision: https://phabricator.sourcevertex.net/D41090