Correct mapping for CTC loss input to reduce exchange
Summary:
The layout in memory is determined by the initial shape the tensor is created with.  Batch needs to be the outermost dimension in order for all timesteps on a tile to be contiguous in memory.  This makes broadcasting/ copying as the function requires more efficient.

Ref T32938

Test Plan: CI

Reviewers: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T32938

Differential Revision: https://phabricator.sourcevertex.net/D41267