Ensure multiple of 4 bytes of input on a tile for CTC Loss
Summary:
With particular plans/input size it is possible to get only a multiple of 2 bytes on a tile, this increases exchange cost.
For this to happen we'd need:
Input data with type half
An odd number of symbols
A set of tiles to partition the input over that results in an odd number of input (num-symbols sized) timeslices per tile. The easiest example is when there are plenty of tiles and we get 1 timestep sized slice per tile.

Ref T32938

Reviewers: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T32938

Differential Revision: https://phabricator.sourcevertex.net/D41527