Simplify CTC inference generate, and merge codelets and partition into copy, extend candidates
Summary:
Split generate codelet into copy, extend candidate generation, and specify which addend to output.  This means that candidates are ready partitioned to select a copy candidate and al extend candidates with a specific addend so that comparison operations are minimised.
Change the merge codelet to compare one copy candidate to many extend candidates which is simpler and allows for multiple parallel codelets to execute.

Ref T31939

Test Plan: Added specific tests, modification to extract the correct codelets to compare for each codelet test

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T31939

Differential Revision: https://phabricator.sourcevertex.net/D42356