Do an explicit copy of the output of the convolution to clone with gather
Summary:
When the output of the convolution is an expression with a slice,
all subsequent element-wise operations on it have as many edges
as regions which are not contiguous. As outputs following
convolution are used typically for element wise operations
(eg optimisers for the GradW, and non-linearity, norms for
Fwd/GradA) the number if edge pointers typically increase.
When there is no slice, the copy should be elided by poplar.

Ref T36156

Test Plan: CI

Reviewers: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: tomm, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T36156

Differential Revision: https://phabricator.sourcevertex.net/D42521