Make aXPlusbY2D/aXPlusbY2D<half,float,false,true> work
Summary:
Trying to use the two vertices `aXPlusbY2D`/`aXPlusbY2D<half,float,false,true>`
gives:
   Unresolved symbol: __runCodelet_popops__aXPlusbY2D___half_float_false_true
This is because there are typos in the assembly source file where the labels
for the entry points for both these vertices are defined.

Also the cycle estimator for the Minus variant is not defined.

Test Plan:
These are currently never tested in the poplibs regression tests, but they
will soon be. As part of T34983, comprehensive testing for all the ScaledAdd
family of vertices will be introduced (this was discovered while developing
that)

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp, Eddy

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, Eddy

Maniphest Tasks: T36817

Differential Revision: https://phabricator.sourcevertex.net/D43393