Adding an extra timestep to the beam history for initial state for CTC Inference
Summary:
In order to remove the need for 2 workarounds in codelets we add an extra timestep to the beam history.
This has an initial state that makes the resulting candidates never merge and so makes the codelet task easier - each initial symbol is an invalid symbol except for the 1st beam which is a voidSymbol.

It has a minimal speed improvement.

Ref T35818

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D43657