CTC Inference Extend beam generation split, and maintain total probability
Summary:
Make the plan struct describe how to partititon the existing stages, including how many vertices to use to make extend candidates.
Use the number of extend candidates vertices to divide work further, in a flexible way.
Take the calculation of total probability away from the select vertex and maintian it from candidate generation onwards.  This makes select do what it says, and allow the total probability calculation to be split between vertices which is faster.  As a downside, there is more data to exchange, and when a merge happens the total probability needs to be re computed with another log add.  The downside appears small though.
Patched up those codelet tests that are still running to check the total probability and the model to find it.

Ref T35818

Test Plan: CI

Reviewers: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: saml, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D43627