Add support for distributed batch norm in poplibs
Summary:
APIs to perform distributed batch norm are similar to
the ones for the non-distributed case except for the
addition of a callback which all-reduce the number
of statistics and gradients where required. The all-reduce
is a reduce-sum rather than reduce-mean due to it being
more efficient at this point.

The callback in the test uses poplibs all-reduce.

Ref T36760

Test Plan: Tests added

Reviewers: Eddy, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, simonl

Reviewed By: Eddy, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, simonl

Subscribers: jaket, alfredb, georgep

Maniphest Tasks: T36760

Differential Revision: https://phabricator.sourcevertex.net/D43579