aXMinusbY2D<half, float> doesn't work when scale values are within tolerance
Summary:
The aXMinusbY2D<half, float> has two code paths, one of which (the fast one) is
chosen if the float scale values can be converted to half without loss of
precision, based on a tolerance value.
When this fast path is chosen, the vertex will incorrectly compute aX + bY
instead aX - bY (because the b constant is not inverted as it should be to use
the code common between the two versions).

Test Plan:
These are currently never tested in the poplibs regression tests, but they
will soon be. As part of T34983, comprehensive testing for all the ScaledAdd
family of vertices will be introduced (this was discovered while developing
that)

Reviewers: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T36946

Differential Revision: https://phabricator.sourcevertex.net/D43404