CTC Inference - split merge codelet by beam, add extra select step
Summary:
Changes to simplify and speed up codelets, doing more work in parallel:

Merge step: There will be beamwidth * beamwidth merge vertices.  Each connected to a single copy candidate (beamwidth copy candidates) and all the extend candidates from a specific beam.
A single comparison is made, based on picking the extend candidate with the same addend as the last symbol output from the beam that the copy candidate came from.  This is the only one that can match.
The copy candidate is modified with the merge result.

A new SelectCopy / SelectExtend step.
This absorbs and parallelises the selection of copy/extend candidates to sort by probability that was in the Select vertex.
There will be beamwidth SelectCopy vertices and beamwidth SelectExtend vertices.  Each loops through beamwidth copy candidates once.
SelectCopy - attached to all broadcast copy candidates from a single beam.  One of these could have been merged, if so it is the output from the vertex, otherwise they are all the same (no merge) and the last one is used
SelectExtend - attached to all copy candidates that were compared to the extend candidates from a specific beam, which are also connected. Check each copy candidate and zero the probability of the extend candidate it was merged with (if any).

The resulting change is that the existing Select vertex is attached to beamwidth copy candidates and all the extend candidates (beamwidth * (numSymbols-1)).  The first stage Select it did do is gone, so now it just does "Select by total probability".  The result is in the first beamwidth candidates (which are connected to the copy candidates, and are so used to Update the beams).

The speed improvement is significant (Note to follow), however the biggest part of work is the same, which is the single threaded, nested loop over all candidates (beamwidth copy candidates + beamwidth * (numClasses-1) extend candidates) in the Select vertex.  This did, and still does dominate.

Ref T35818

Test Plan: CI

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T35818

Differential Revision: https://phabricator.sourcevertex.net/D43810