Fix cycle count regression in internal SDK tests
Summary:
The output shape of the tensor after the `reduceWithOutputProgOrCss`
call was in column-major order, and so required additional rearrangement
to then use, which caused a cycle-count regression in some of the internal
SDK tests.

The column-major order was due to an optimisation in `reduceFirstDim2D`
writing the column-major slices back to the `out` variable, which should not
be modified by this function in the case where it has been provided by the
user.

This patch changes the `reduceFirstDim2D` function so that it preserves
the original value of the `out` tensor if it has a user-provided value.

This also fixes a use-after-move bug when constructing some debug information.

Test Plan:
Ran:
```
pytest -s -v synthetic_benchmarks/popart/transformer/test_popart_transformer_performance.py::TestInfer::test_ff_128_svl_1024_tvl_1024_shards_1_no_grouped_attention
```
and observed that with this patch the cycles returned to the same number as with SDK 528 (356754139).
Without this patch the cycle count is 372654139 cycles, which is worse than just running the reductions
serially.

Also CI.

Reviewers: godfrey.da.costa, richard, samk, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, timp

Reviewed By: godfrey.da.costa, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T37801

Differential Revision: https://phabricator.sourcevertex.net/D44232