Strided reduce with outer stride, extending the "regular partials" description of inputs to reductions
Summary:
Introduced an outerStride, and numOuterStrides to the description of inputs held in regularPartials, and used in strided reduce.
An outerStride describes a complete, identical repetition of what was implemented before (an inner pattern) so, where inner patterns can be:
```
Numbers represent partials elements that belong to a specific column, xx are elements that aren't part of the intended reduction
00112233xx
00112233xx
00112233xx
```
Where the regular partials struct would have
output with 8 elements (All instances 00112233 reduced to one instance)
innerFactor =  2 (Each column is repeated twice)
outerFactor = 3 (00112233xx is repeated 3x)
stride = 10 (The pattern 00112233xx has 10 elements)

We now allow multiple repetitions of the same thing:

```
00112233xx
00112233xx
00112233xx
xxxxxxxx
xxxxxxxx
00112233xx
00112233xx
00112233xx
```
numOuterStrides = 2 (2 repetitions of the whole pattern)
outerStride = elements between 1st 0 in the repetitions of the whole pattern

outerStride is passed to the vertex in a form that simplifies the implementation

Fixes T36913

Test Plan:
Verified that it does what it should.
Checking that the change to describe regular partials in a different way doesn't break anything where they aren't needed is difficult.
Added many more random tests.  Checked that many of the paths through the code are exercised.

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, tomm

Maniphest Tasks: T36913

Differential Revision: https://phabricator.sourcevertex.net/D43456