CTC Inference Ranking method of Sorting candidates
Summary:
The main purpose of this change is to add the ranking method of sorting to the code.  This is an option to the existing Select method because Rank is faster only if enough worker threads are available.

Rank:
Copy all candidates to sort to all tiles.
Use a Multivertex to run 6 workers at once with a single output of size beamwidth per tile, for any items that the vertex finds that are ranked in the most likely `beamwidth` candidates.
Each worker is assigned a unique candidate to Rank.
It counts how many candidates have a higher probability than the one it is assigned and if ranked in the top `beamwidth` it writes into that vertex's output.

This makes candidatesToSort / 6 results which are then reduced in a second step.

The rank, and reduce "operations" are simple.

In doing this I found the need to add options to test it - to choose the sort method, and to force half/float partials.
So half/float partials is debugged and a test added.
Also I was interested in how things grew with large numbers of symbols.  This is quickly limited unless partitions for sorting and generating extend candidates can contain more candidates.  In other words if we need numSymbols-1 partitions to generate candidates, if numSymbols > tiles it won't fit.  So this diff contains some changes to make that more flexible which the "planner" pushes in if necessary.

Ref T35818
Ref T31939

Test Plan: CI, Tests to check options that are passed and where tiles constrain the "plan"

Reviewers: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Reviewed By: #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, saml

Maniphest Tasks: T35818, T31939

Differential Revision: https://phabricator.sourcevertex.net/D44073